define common dataframe api for DataFrameSchema
